requirements.txt
appdirs==1.4.3
boto3==1.9.210
botocore==1.12.212
docutils==0.14
Flask==1.1.1
Flask-Cors==3.0.3
Flask-HTTPAuth==3.3.0
Flask-Login==0.4.1
Jinja2==2.11
jsonschema==2.6.0
line-bot-sdk==1.12.1
numpy==1.20
opencv-contrib-python==4.5.1.48
opencv-python==4.5.1.48
opencv-python-headless==4.5.1.48
openpyxl==2.4.10
path.py==10.5
pathlib2==2.3.0
Pillow>=6.2.0
protobuf==3.7.1
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
psycopg2==2.8.6
PySocks==1.6.7
python-dateutil==2.6.1
python-dotenv==0.10.3
pytz==2017.3
PyWavelets==1.1.1
requests==2.22.0
requests-oauthlib==1.2.0
requests-toolbelt==0.9.1
Werkzeug== 0.16.0 
wheel==0.36.2

